requirements.txt
accelerate==0.18.0
altgraph==0.17.3
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
cmake==3.27.2
comm==0.1.2
cryptography==38.0.4
defusedxml==0.7.1
denoising-diffusion-pytorch==1.5.4
diffusers==0.18.2
einops==0.6.1
ema-pytorch==0.2.3
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.12.2
fire==0.5.0
flit_core==3.6.0
future==0.18.3
idna==3.4
imageio==2.28.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.6
jsonschema==4.16.0
jupyter==1.0.0
jupyter_client==7.4.8
jupyter-console==6.4.4
jupyter_core==5.1.1
jupyter-server==1.23.4
jupyterlab==3.5.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.16.5
jupyterlab-widgets==1.0.0
lazy_loader==0.2
lit==16.0.6
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==0.8.4
mpmath==1.3.0
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
notebook_shim==0.2.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==22.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==22.3.1
platformdirs==2.5.2
ply==3.11
prometheus-client==0.14.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarmor==8.3.3
pyarmor.cli.core==4.3.1
pycparser==2.21
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-sip==12.12.2
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-fid==0.3.0
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
qtconsole==5.4.0
QtPy==2.2.0
scikit-image==0.20.0
scipy==1.11.1
Send2Trash==1.8.0
sip==6.7.11
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
sympy==1.12
termcolor==2.2.0
terminado==0.17.1
tifffile==2023.4.12
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
transformers==4.29.2
tqdm==4.66.1
traitlets==5.7.1
triton==2.0.0
urllib3==1.26.14
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.37.1
widgetsnbextension==3.5.2

